+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -1,9 +1,9 @@
-#include <stdio.h>
-#include <string.h>
 #include <stdlib.h>
+#include <string.h>
+#include <stdio.h>
 #include "lists.h"
 
-/**
+/*
  * listint_len - return The number of nodes in the list.
  * @h: The listint_t list.
  *
@@ -12,12 +12,12 @@
 
 size_t listint_len(const listint_t *h)
 {
-    int nodes = 0;
+	size_t nodes = 0;
 
-    while (h)
-    {
-        h = h->next;
-        nodes++;
-    }
-    return (nodes);
+	while (h)
+	{
+		h = h->next;
+		nodes++;
+	}
+	return (nodes);
 }
